extract EntrySourcePositions
